Move payjoin v2 integration tests to `payjoin-defaults`
