+++ payjoin-defaults/Cargo.toml
@@ -20,4 +20,18 @@ rustls = { version = "0.22.2", optional = true }
 ureq = "2.9.4"
 
 [dev-dependencies]
-
+bitcoin = { version = "0.30.0", features = ["base64"] }
+bitcoind = { version = "0.31.1", features = ["0_21_2"] }
+http = "1"
+log = { version = "0.4.14"}
+ohttp-relay = "0.0.7"
+once_cell = "1"
+payjoin-directory = { path = "../payjoin-directory", features = ["danger-local-https"] }
+rcgen = { version = "0.11" }
+rustls = "0.22.2"
+testcontainers = "0.15.0"
+testcontainers-modules = { version = "0.1.3", features = ["redis"] }
+tokio = { version = "1.12.0", features = ["full"] }
+tracing = "0.1.40"
+tracing-subscriber = { version = "0.3.17", features = ["env-filter"] }
+url = "2.2.2"
